refactor: move local component
