+++ src/shared/components/Drawer/components/Header/HeaderView.tsx
@@ -5,7 +5,7 @@ import { FiUser } from "react-icons/fi";
 import { Link } from "react-router-dom";
 import Logo from "shared/components/Logo";
 import { ICON_SIZE } from "shared/constants";
-import Dropdown from "./Dropdown";
+import Dropdown from "shared/components/Drawer/components/Header/components/Dropdown";
 import { Header, Nav } from "./style";
 import { IHeaderView } from "./types/Header.component";
