+++ src/shared/components/Drawer/components/Header/components/Dropdown/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./dropdown.types";
